requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
djangorestframework==3.14.0
redis==4.6.0
django-redis==5.3.0
googletrans==3.1.0a0